requirements.txt
Django
djangorestframework
requests
pygetwindow
mysqlclient
